refactoring local grpc to interop
