+++ openfl-workspace/flower-app-pytorch/src/grpc/connector/__init__.py
@@ -1,1 +1,1 @@
-from src.grpc.connector.utils import get_local_grpc_server
+from src.grpc.connector.utils import get_interop_server
