+++ openfl-workspace/flower-app-pytorch/src/grpc/connector/utils.py
@@ -1,9 +1,9 @@
 import importlib
 
-def get_local_grpc_server(framework: str = 'Flower') -> object:
+def get_interop_server(framework: str = 'Flower') -> object:
     if framework == 'Flower':
         try:
-            module = importlib.import_module('src.grpc.connector.flower.local_grpc_server')
+            module = importlib.import_module('src.grpc.connector.flower.interop_server')
             return module.FlowerInteropServer
         except ImportError:
             print("Flower is not installed.")
